Extract ExecuteAction wire protocol types into backend/protocol.ts
