+++ packages/plugins/apps/src/backend/client/types.ts
@@ -3,29 +3,6 @@
 // Copyright 2019-Present Datadog, Inc.
 
 /**
- * Request payload for executing a backend function
- */
-export interface ExecuteActionRequest {
-    functionName: string;
-    args: unknown[];
-}
-
-/**
- * Response from executing a backend function
- */
-export type ExecuteActionResponse<TData = unknown> =
-    | {
-          success: true;
-          result: {
-              data: TData;
-          };
-      }
-    | {
-          success: false;
-          error: string;
-      };
-
-/**
  * A transport function that executes a backend function via a specific mechanism.
  */
 export type BackendFunctionTransport = <TData>(
